requirements.txt
asgiref==3.8.1
channels==4.2.2
channels_redis==4.2.1
chardet==5.2.0
Django==5.1.7
django-cors-headers==4.7.0
django-filter==25.1
django-restframework==0.0.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
msgpack==1.1.0
pillow==11.1.0
psycopg2-binary==2.9.10
PyJWT==2.9.0
PyPDF2==3.0.1
python-dotenv==1.1.0
redis==5.2.1
reportlab==4.3.1
sqlparse==0.5.3
tzdata==2025.1
